+++ src/module_library/c3_leaf_photosynthesis.cpp
@@ -1,6 +1,7 @@
 #include "c3_leaf_photosynthesis.h"
-#include "c3photo.h"              // for c3photoC
-#include "leaf_energy_balance.h"  // for leaf_energy_balance
+#include "c3photo.h"                  // for c3photoC
+#include "leaf_energy_balance.h"      // for leaf_energy_balance
+#include "c3_temperature_response.h"  // for c3_temperature_response_parameters
 
 using standardBML::c3_leaf_photosynthesis;
 
@@ -19,16 +20,37 @@ string_vector c3_leaf_photosynthesis::get_inputs()
         "electrons_per_oxygenation",    // electron / oxygenation
         "gbw_canopy",                   // m / s
         "Gs_min",                       // mol / m^2 / s
+        "Gstar_c",                      // dimensionless
+        "Gstar_Ea",                     // J / mol
         "height",                       // m
         "jmax",                         // micromol / m^2 / s
+        "Jmax_c",                       // dimensionless
+        "Jmax_Ea",                      // J / mol
+        "Kc_c",                         // dimensionless
+        "Kc_Ea",                        // J / mol
+        "Ko_c",                         // dimensionless
+        "Ko_Ea",                        // J / mol
         "leafwidth",                    // m
         "O2",                           // mmol / mol
+        "phi_PSII_0",                   // dimensionless
+        "phi_PSII_1",                   // (degrees C)^(-1)
+        "phi_PSII_2",                   // (degrees C)^(-2)
         "Rd",                           // micromol / m^2 / s
+        "Rd_c",                         // dimensionless
+        "Rd_Ea",                        // J / mol
         "rh",                           // dimensionless
         "StomataWS",                    // dimensionless
         "temp",                         // degrees C
-        "theta",                        // dimensionless
+        "theta_0",                      // dimensionless
+        "theta_1",                      // (degrees C)^(-1)
+        "theta_2",                      // (degrees C)^(-2)
+        "Tp_c",                         // dimensionless
+        "Tp_Ha",                        // J / mol
+        "Tp_Hd",                        // J / mol
+        "Tp_S",                         // J / K / mol
         "tpu_rate_max",                 // micromol / m^2 / s
+        "Vcmax_c",                      // dimensionless
+        "Vcmax_Ea",                     // J / mol
         "vmax1",                        // micromol / m^2 / s
         "windspeed"                     // m / s
     };
@@ -55,6 +77,31 @@ string_vector c3_leaf_photosynthesis::get_outputs()
 
 void c3_leaf_photosynthesis::do_operation() const
 {
+    // Combine temperature response parameters
+    c3_temperature_response_parameters const tr_param{
+        Gstar_c,
+        Gstar_Ea,
+        Jmax_c,
+        Jmax_Ea,
+        Kc_c,
+        Kc_Ea,
+        Ko_c,
+        Ko_Ea,
+        phi_PSII_0,
+        phi_PSII_1,
+        phi_PSII_2,
+        Rd_c,
+        Rd_Ea,
+        theta_0,
+        theta_1,
+        theta_2,
+        Tp_c,
+        Tp_Ha,
+        Tp_Hd,
+        Tp_S,
+        Vcmax_c,
+        Vcmax_Ea};
+
     // Make an initial guess for boundary layer conductance
     double const gbw_guess{1.2};  // mol / m^2 / s
 
@@ -62,9 +109,9 @@ void c3_leaf_photosynthesis::do_operation() const
     // air temperature
     double const initial_stomatal_conductance =
         c3photoC(
-            absorbed_ppfd, ambient_temperature, ambient_temperature,
+            tr_param, absorbed_ppfd, ambient_temperature, ambient_temperature,
             rh, vmax1, jmax, tpu_rate_max, Rd, b0,
-            b1, Gs_min, Catm, atmospheric_pressure, O2, theta, StomataWS,
+            b1, Gs_min, Catm, atmospheric_pressure, O2, StomataWS,
             electrons_per_carboxylation,
             electrons_per_oxygenation, beta_PSII, gbw_guess)
             .Gs;  // mol / m^2 / s
@@ -88,10 +135,10 @@ void c3_leaf_photosynthesis::do_operation() const
     // using the new leaf temperature
     const photosynthesis_outputs photo =
         c3photoC(
-            absorbed_ppfd, leaf_temperature, ambient_temperature,
+            tr_param, absorbed_ppfd, leaf_temperature, ambient_temperature,
             rh, vmax1, jmax,
             tpu_rate_max, Rd, b0, b1, Gs_min, Catm, atmospheric_pressure, O2,
-            theta, StomataWS,
+            StomataWS,
             electrons_per_carboxylation, electrons_per_oxygenation, beta_PSII,
             et.boundary_layer_conductance);